+++ pkg/scheme/core/v1/registry.go
@@ -89,37 +89,8 @@ func addKnownTypes(scheme *runtime.Scheme) error {
 		&ConfigMapList{},
 		&CloudProvider{},
 		&CloudProviderList{},
+		&Registry{},
+		&RegistryList{},
 	)
 	return nil
-}
-
-// +genclient
-// +genclient:nonNamespaced
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-// +k8s:openapi-gen=false
-
-type Registry struct {
-	metav1.TypeMeta `json:",inline"`
-	// Standard object's metadata.
-	// +optional
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	RegistrySpec      `json:",inline"`
-}
-
-type RegistrySpec struct {
-	Scheme     string `json:"scheme,omitempty"`
-	Host       string `json:"host,omitempty"`
-	SkipVerify bool   `json:"skipVerify,omitempty"`
-	CA         string `json:"ca,omitempty"`
-}
-
-// RegistryList is a resource containing a list of RegistryList objects.
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-type RegistryList struct {
-	metav1.TypeMeta
-	// +optional
-	metav1.ListMeta
-
-	// Items is the list of registry.
-	Items []Registry
 }
